Pull out common test base
