+++ source/Calamari.Tests/KubernetesFixtures/Authentication/SetupKubectlAuthenticationAwsFixture.cs
@@ -1,24 +1,16 @@
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using Calamari.Aws.Deployment;
-using Calamari.Common.Features.Processes;
-using Calamari.Common.Plumbing.FileSystem;
-using Calamari.Common.Plumbing.Logging;
-using Calamari.Common.Plumbing.Variables;
 using Calamari.Kubernetes;
-using Calamari.Kubernetes.Integration;
+using Calamari.Kubernetes.Authentication;
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
 
 namespace Calamari.Tests.KubernetesFixtures.Authentication
 {
-    public class SetupKubectlAuthenticationAwsFixture
+    public class SetupKubectlAuthenticationAwsFixture : BaseSetupKubectlAuthenticationFixture
     {
-        private readonly string workingDirectory = Path.Combine("working", "directory");
-        private const string Namespace = "my-cool-namespace";
-
         private const string CurrentAwsVersion = "aws-cli/2.14.2";
         private const string OlderAwsVersion = "aws-cli/1.16.155";
         private const string InvalidAwsVersion = "aws-cli/not-a-version";
@@ -28,68 +20,19 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
         private const string AwsClusterUrl = "http://www." + AwsRegion + ".eks.amazonaws.com";
         private const string InvalidAwsClusterUrl = "http://www." + AwsRegion + "..eks.amazonaws.com";
 
-        private IVariables variables;
-        private ILog log;
-        private ICommandLineRunner commandLineRunner;
-        private IKubectl kubectl;
-        private ICalamariFileSystem fileSystem;
-        private Dictionary<string, string> environmentVars;
-
-        private SetupKubectlAuthenticationFixture.Invocations invocations;
-
-        // TODO: Strip down/commonize
         [SetUp]
         public void Setup()
         {
-            invocations = new SetupKubectlAuthenticationFixture.Invocations();
-            invocations.AddLogMessageFor("which", "kubelogin", "kubelogin");
-            invocations.AddLogMessageFor("where", "kubelogin", "kubelogin");
             AddLogForAwsVersion(CurrentAwsVersion);
-
-            variables = new CalamariVariables();
-
-            log = Substitute.For<ILog>();
-            commandLineRunner = Substitute.For<ICommandLineRunner>();
-            commandLineRunner.Execute(Arg.Any<CommandLineInvocation>())
-                .Returns(
-                    x =>
-                    {
-                        var invocation = x.Arg<CommandLineInvocation>();
-                        var isSuccess = true;
-                        string logMessage = null;
-                        if (invocation.Executable != "chmod")
-                        {
-                            isSuccess = invocations.TryAdd(invocation.Executable, invocation.Arguments, out logMessage);
-                        }
-                        if (logMessage != null)
-                            invocation.AdditionalInvocationOutputSink?.WriteInfo(logMessage);
-                        return new CommandResult(
-                            invocation.Executable,
-                            isSuccess ? 0 : 1,
-                            workingDirectory: workingDirectory);
-                    });
-            kubectl = Substitute.For<IKubectl>();
-            kubectl.ExecutableLocation.Returns("kubectl");
-            kubectl.When(x => x.ExecuteCommandAndAssertSuccess(Arg.Any<string[]>()))
-                .Do(
-                    x =>
-                    {
-                        var args = x.Arg<string[]>();
-                        if (args != null)
-                            invocations.TryAdd("kubectl", string.Join(" ", args), out var _);
-                    });
-            fileSystem = Substitute.For<ICalamariFileSystem>();
 
             variables.Set(SpecialVariables.ClusterUrl, AwsClusterUrl);
             variables.Set(SpecialVariables.EksClusterName, EksClusterName);
             variables.Set(SpecialVariables.Namespace, Namespace);
+            variables.Set(Deployment.SpecialVariables.Account.AccountType, AccountTypes.AmazonWebServicesAccount);
 
-            environmentVars = new Dictionary<string, string>
-            {
-                { "AWS_ACCESS_KEY_ID", "access_key" },
-                { "AWS_SECRET_ACCESS_KEY", "secret_key" },
-                { "AWS_REGION", "region" }
-            };
+            environmentVars.Add("AWS_ACCESS_KEY_ID", "access_key");
+            environmentVars.Add("AWS_SECRET_ACCESS_KEY", "secret_key");
+            environmentVars.Add("AWS_REGION", "region");
         }
 
         SetupKubectlAuthentication CreateSut() =>
@@ -132,6 +75,7 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
             AddLogForWhichAws();
             AddLogForAwsVersion(CurrentAwsVersion);
             variables.AddFlag(AwsSpecialVariables.Authentication.UseInstanceRole, true);
+            variables.Set(Deployment.SpecialVariables.Account.AccountType, "");
 
             var expectedInvocations = SetupClusterContextInvocations(AwsClusterUrl);
             expectedInvocations.AddRange(AwsCliInvocations());
@@ -233,9 +177,8 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
                     GetNamespaceInvocation
                 });
 
-            var result = CreateSut().Execute();
+            CreateSut().Execute();
 
-            result.VerifySuccess();
             AssertInvocations(expectedInvocations);
             log.Received()
                 .Verbose(
@@ -278,13 +221,13 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
         {
             invocations.AddLogMessageFor(
                 "aws",
-                $"eks get-token --cluster-name={EksClusterName} --region={AwsRegion}",
+                $"eks get-token --cluster-name={EksClusterName} --region={AwsRegion} --profile octopus",
                 $"{{ \"apiVersion\": \"1.2.3\"}}");
         }
 
         void AddLogForAwsVersion(string version)
         {
-            invocations.AddLogMessageFor("aws", "--version", version);
+            invocations.AddLogMessageFor("aws", "--version --profile octopus", version);
         }
 
         List<(string, string)> SetupClusterContextInvocations(string clusterUser)
@@ -302,11 +245,11 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
         {
             return new List<(string, string)>
             {
-                ("aws", "configure set aws_access_key_id access_key"),
-                ("aws", "configure set aws_secret_access_key secret_key"),
-                ("aws", "configure set aws_default_region region"),
-                ("aws", "configure set aws_session_token"),
-                ("aws", "--version")
+                ("aws", "configure set aws_access_key_id access_key --profile octopus"),
+                ("aws", "configure set aws_secret_access_key secret_key --profile octopus"),
+                ("aws", "configure set aws_default_region region --profile octopus"),
+                ("aws", "configure set aws_session_token --profile octopus"),
+                ("aws", "--version --profile octopus")
             };
         }
 
@@ -315,14 +258,14 @@ namespace Calamari.Tests.KubernetesFixtures.Authentication
                 $"config set-credentials octouser --exec-command=aws-iam-authenticator --exec-api-version=client.authentication.k8s.io/v1alpha1 --exec-arg=token --exec-arg=-i --exec-arg={EksClusterName}");
 
         (string, string ) GetAwsTokenInvocation =>
-            ("aws", "eks get-token --cluster-name=my-cool-eks-cluster-name --region=southwest");
+            ("aws", "eks get-token --cluster-name=my-cool-eks-cluster-name --region=southwest --profile octopus");
 
         (string, string) SetKubectlCredentialsWithAwsCliInvocation =>
             ("kubectl",
-                "config set-credentials octouser --exec-command=aws --exec-arg=eks --exec-arg=get-token --exec-arg=--cluster-name=my-cool-eks-cluster-name --exec-arg=--region=southwest --exec-api-version=1.2.3");
+                "config set-credentials octouser --exec-command=aws --exec-arg=eks --exec-arg=get-token --exec-arg=--cluster-name=my-cool-eks-cluster-name --exec-arg=--region=southwest --exec-api-version=1.2.3 --exec-env AWS_PROFILE=octopus");
 
         (string, string) GetNamespaceInvocation =>
-            ("kubectl", $"get namespace {Namespace} --request-timeout=1m");
+            ("kubectl", $"get namespace {Namespace}");
 
         private void AssertInvocations(List<(string, string)> expectedInvocations)
         {
